3 Lesson

+++ OOP_basics.Tests/BankAccountTests.cs
@@ -13,7 +13,8 @@ namespace OOP_basics.Tests
             BankAccount _bankAccount = new BankAccount(100, AccountType.Current);
             int expected = 200;
             _bankAccount.PutMoney(100);
-            int actual = _bankAccount.Balance;
+            double actual = _bankAccount.Balance;
+
 
             Assert.Equal(expected, actual);
         }
@@ -47,7 +48,7 @@ namespace OOP_basics.Tests
             BankAccount _bankAccount = new BankAccount(100, AccountType.Current);
             int expected = 90;
             _bankAccount.Withdrawmoney(10);
-            int actual = _bankAccount.Balance;
+            double actual = _bankAccount.Balance;
 
             Assert.Equal(expected, actual);
         }
